refactor execution.Env to move into runtime package
