+++ internal/core/execution/context.go
@@ -52,6 +52,11 @@ func (e DAGContext) UserEnvsMap() map[string]string {
 	return result
 }
 
+// DAGRunRef returns the DAGRunRef for the current DAG context.
+func (e DAGContext) DAGRunRef() DAGRunRef {
+	return NewDAGRunRef(e.DAG.Name, e.DAGRunID)
+}
+
 // AllEnvs returns all environment variables as a slice of strings in "key=value" format.
 // Includes OS environment (BaseEnv). Use this for command executor and DAG runner.
 // Secrets have the highest priority and are appended last.
@@ -161,6 +166,12 @@ func SetupDAGContext(ctx context.Context, dag *core.DAG, db Database, rootDAGRun
 		CoordinatorCli: coordinatorCli,
 		Shell:          dag.Shell,
 	})
+}
+
+// WithDAGContext returns a new context with the given DAGContext.
+// This is useful for tests that need to set up a DAGContext directly.
+func WithDAGContext(ctx context.Context, dagCtx DAGContext) context.Context {
+	return context.WithValue(ctx, dagCtxKey{}, dagCtx)
 }
 
 // GetDAGContext retrieves the DAGContext from the context.
