+++ internal/runtime/debug_test.go
@@ -10,16 +10,19 @@ import (
 	"github.com/dagu-org/dagu/internal/runtime"
 )
 
+// Note: This test uses execution.SetupDAGContext to set up the DAG context,
+// then uses runtime.GetEnv/WithEnv to manage the runtime Env.
+
 func TestDebugVariables(t *testing.T) {
 	// Create a test context with environment variables
 	ctx := execution.SetupDAGContext(context.Background(), &core.DAG{}, nil, execution.DAGRunRef{}, "test-run", "test.log", nil, nil, nil)
-	env := execution.GetEnv(ctx)
+	env := runtime.GetEnv(ctx)
 
 	// Store variable with spaces
 	env.Variables.Store("SPACES", "SPACES=  ")
 
 	// IMPORTANT: Update the context with the modified env
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	// Print out what Variables() returns
 	vars := env.Variables.Variables()
@@ -43,14 +46,14 @@ func TestDebugVariables(t *testing.T) {
 	fmt.Printf("\nEnvs map: %#v\n", env.Envs)
 
 	// Let's check what GetEnv returns
-	envFromCtx := execution.GetEnv(ctx)
+	envFromCtx := runtime.GetEnv(ctx)
 	fmt.Printf("\nEnv from context Variables: %#v\n", envFromCtx.Variables.Variables())
 
 	// Test with special characters too
 	env.Variables.Store("SPECIAL", "SPECIAL=$pecial!@#")
 
 	// Update context again after adding new variable
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	vars2 := env.Variables.Variables()
 	fmt.Printf("\nVariables map with special: %#v\n", vars2)
@@ -67,7 +70,7 @@ func TestDebugVariables(t *testing.T) {
 	env.Variables.Store("DOLLAR", "DOLLAR=$")
 	env.Variables.Store("DOLLAR_P", "DOLLAR_P=$p")
 	env.Variables.Store("ESCAPED", "ESCAPED=\\$pecial")
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	r3, _ := runtime.EvalString(ctx, "${DOLLAR}")
 	r4, _ := runtime.EvalString(ctx, "${DOLLAR_P}")
@@ -80,7 +83,7 @@ func TestDebugVariables(t *testing.T) {
 	// Test without environment expansion
 	fmt.Printf("\n--- Testing without env expansion ---\n")
 	env.Variables.Store("TEST_DOLLAR", "TEST_DOLLAR=$HOME/test")
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	// This should expand $HOME
 	r6, _ := runtime.EvalString(ctx, "${TEST_DOLLAR}")
