+++ internal/runtime/eval.go
@@ -8,24 +8,23 @@ import (
 
 	"github.com/dagu-org/dagu/internal/common/cmdutil"
 	"github.com/dagu-org/dagu/internal/common/stringutil"
-	"github.com/dagu-org/dagu/internal/core/execution"
 )
 
 // EvalString evaluates the given string with the variables within the execution context.
 func EvalString(ctx context.Context, s string, opts ...cmdutil.EvalOption) (string, error) {
-	return execution.GetEnv(ctx).EvalString(ctx, s, opts...)
+	return GetEnv(ctx).EvalString(ctx, s, opts...)
 }
 
 // EvalBool evaluates the given value with the variables within the execution context
 // and parses it as a boolean.
 func EvalBool(ctx context.Context, value any) (bool, error) {
-	return execution.GetEnv(ctx).EvalBool(ctx, value)
+	return GetEnv(ctx).EvalBool(ctx, value)
 }
 
 // EvalObject recursively evaluates the string fields of the given object
 // with the variables within the execution context.
 func EvalObject[T any](ctx context.Context, obj T) (T, error) {
-	vars := execution.GetEnv(ctx).VariablesMap()
+	vars := GetEnv(ctx).VariablesMap()
 	return cmdutil.EvalObject(ctx, obj, vars)
 }
 
@@ -39,10 +38,10 @@ func GenerateSubDAGRunID(ctx context.Context, params string, repeated bool) stri
 			randomBytes = []byte(fmt.Sprintf("%d", time.Now().UnixNano()))
 		}
 		return stringutil.Base58EncodeSHA256(
-			fmt.Sprintf("%s:%s:%s:%x", execution.GetEnv(ctx).DAGRunID, execution.GetEnv(ctx).Step.Name, params, randomBytes),
+			fmt.Sprintf("%s:%s:%s:%x", GetEnv(ctx).DAGRunID, GetEnv(ctx).Step.Name, params, randomBytes),
 		)
 	}
-	env := execution.GetEnv(ctx)
+	env := GetEnv(ctx)
 	return stringutil.Base58EncodeSHA256(
 		fmt.Sprintf("%s:%s:%s", env.DAGRunID, env.Step.Name, params),
 	)
