+++ internal/runtime/eval_test.go
@@ -10,15 +10,18 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+// Note: These tests use execution.SetupDAGContext to set up the DAG context,
+// then use runtime.NewEnvForStep/GetEnv/WithEnv to manage the runtime Env.
+
 func TestEvalString(t *testing.T) {
 	t.Parallel()
 
 	// Create a test context with environment variables
 	ctx := context.Background()
-	env := execution.NewEnv(ctx, core.Step{Name: "test-step"})
+	env := runtime.NewEnvForStep(ctx, core.Step{Name: "test-step"})
 	env.Variables.Store("TEST_VAR", "TEST_VAR=hello")
 	env.Envs["ANOTHER_VAR"] = "world"
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	tests := []struct {
 		name     string
@@ -64,13 +67,13 @@ func TestEvalString(t *testing.T) {
 func TestEvalBool(t *testing.T) {
 	// Create a test context with environment variables
 	ctx := context.Background()
-	env := execution.NewEnv(ctx, core.Step{Name: "test-step"})
+	env := runtime.NewEnvForStep(ctx, core.Step{Name: "test-step"})
 	env.Variables.Store("TRUE_VAR", "TRUE_VAR=true")
 	env.Variables.Store("FALSE_VAR", "FALSE_VAR=false")
 	env.Variables.Store("ONE_VAR", "ONE_VAR=1")
 	env.Variables.Store("ZERO_VAR", "ZERO_VAR=0")
 	env.Variables.Store("INVALID_VAR", "INVALID_VAR=not-a-bool")
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	tests := []struct {
 		name     string
@@ -156,11 +159,11 @@ type NestedStruct struct {
 func TestEvalObject(t *testing.T) {
 	// Create a test context with environment variables
 	ctx := context.Background()
-	env := execution.NewEnv(ctx, core.Step{Name: "test-step"})
+	env := runtime.NewEnvForStep(ctx, core.Step{Name: "test-step"})
 	env.Variables.Store("NAME_VAR", "NAME_VAR=John")
 	env.Variables.Store("DESC_VAR", "DESC_VAR=Developer")
 	env.Variables.Store("NESTED_VAR", "NESTED_VAR=NestedValue")
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	// Create a test struct
 	testObj := TestStruct{
@@ -198,11 +201,11 @@ func TestEvalObjectWithExecutorConfig(t *testing.T) {
 
 	// Create a test context with environment variables
 	ctx := context.Background()
-	env := execution.NewEnv(ctx, core.Step{Name: "test-step"})
+	env := runtime.NewEnvForStep(ctx, core.Step{Name: "test-step"})
 	env.Variables.Store("EXECUTOR_TYPE", "EXECUTOR_TYPE=docker")
 	env.Variables.Store("HOST_VAR", "HOST_VAR=localhost")
 	env.Variables.Store("PORT_VAR", "PORT_VAR=8080")
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	// Create an ExecutorConfig with variables
 	config := core.ExecutorConfig{
@@ -251,10 +254,10 @@ func TestGenerateSubDAGRunID(t *testing.T) {
 
 	// Create a test context with environment variables
 	ctx := context.Background()
-	env := execution.NewEnv(ctx, core.Step{Name: "test-step"})
+	env := runtime.NewEnvForStep(ctx, core.Step{Name: "test-step"})
 	env.DAGRunID = "parent-run-123"
 	env.Step.Name = "child-step"
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	tests := []struct {
 		name      string
@@ -307,11 +310,11 @@ func TestEvalObjectWithComplexNestedStructures(t *testing.T) {
 
 	// Create a test context with environment variables
 	ctx := context.Background()
-	env := execution.NewEnv(ctx, core.Step{Name: "test-step"})
+	env := runtime.NewEnvForStep(ctx, core.Step{Name: "test-step"})
 	env.Variables.Store("VAR1", "VAR1=value1")
 	env.Variables.Store("VAR2", "VAR2=value2")
 	env.Variables.Store("NUM", "NUM=42")
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	tests := []struct {
 		name     string
@@ -467,11 +470,11 @@ func TestEvalStringEdgeCases(t *testing.T) {
 
 	// Create a test context with environment variables
 	ctx := execution.SetupDAGContext(context.Background(), &core.DAG{}, nil, execution.DAGRunRef{}, "test-run", "test.log", nil, nil, nil)
-	env := execution.GetEnv(ctx)
+	env := runtime.GetEnv(ctx)
 	env.Variables.Store("EMPTY", "EMPTY=")
 	env.Variables.Store("SPACES", "SPACES=  ")
 	env.Variables.Store("SPECIAL", "SPECIAL=special!@#")
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	tests := []struct {
 		name     string
@@ -533,11 +536,11 @@ func TestEvalStringEdgeCases(t *testing.T) {
 func TestEvalObjectWithDirectStringEvaluation(t *testing.T) {
 	// Create a test context with environment variables
 	ctx := context.Background()
-	env := execution.NewEnv(ctx, core.Step{Name: "test-step"})
+	env := runtime.NewEnvForStep(ctx, core.Step{Name: "test-step"})
 	env.Variables.Store("STRING_VAR", "STRING_VAR=evaluated_string")
 	env.Variables.Store("PATH_VAR", "PATH_VAR=/path/to/file")
 	env.Variables.Store("COMBINED", "COMBINED=prefix")
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	tests := []struct {
 		name     string
@@ -618,7 +621,7 @@ func TestEvalBoolEdgeCases(t *testing.T) {
 	// Create a test context with environment variables
 	ctx := execution.SetupDAGContext(context.Background(), &core.DAG{}, nil, execution.DAGRunRef{}, "test-run", "test.log", nil, nil, nil)
 
-	env := execution.GetEnv(ctx)
+	env := runtime.GetEnv(ctx)
 	env.Variables.Store("YES", "YES=yes")
 	env.Variables.Store("NO", "NO=no")
 	env.Variables.Store("ON", "ON=on")
@@ -626,7 +629,7 @@ func TestEvalBoolEdgeCases(t *testing.T) {
 	env.Variables.Store("T", "T=t")
 	env.Variables.Store("F", "F=f")
 
-	ctx = execution.WithEnv(ctx, env)
+	ctx = runtime.WithEnv(ctx, env)
 
 	tests := []struct {
 		name     string
