+++ internal/runtime/runner.go
@@ -417,7 +417,7 @@ func (r *Runner) teardownNode(node *Node) error {
 }
 
 func (r *Runner) setupVariables(ctx context.Context, plan *Plan, node *Node) context.Context {
-	env := NewEnv(ctx, node.Step(), plan)
+	env := NewPlanEnv(ctx, node.Step(), plan)
 
 	// Load output variables from predecessor nodes (dependencies)
 	// This traverses backwards from the current node to find all nodes it depends on
@@ -473,11 +473,11 @@ func (r *Runner) setupVariables(ctx context.Context, plan *Plan, node *Node) con
 
 	env.ForceLoadOutputVariables(envVars)
 
-	return execution.WithEnv(ctx, env)
+	return WithEnv(ctx, env)
 }
 
 func (r *Runner) setupEnvironEventHandler(ctx context.Context, plan *Plan, node *Node) context.Context {
-	env := NewEnv(ctx, node.Step(), plan)
+	env := NewPlanEnv(ctx, node.Step(), plan)
 	env.Envs[execution.EnvKeyDAGRunStatus] = r.Status(ctx, plan).String()
 
 	// get all output variables
@@ -489,7 +489,7 @@ func (r *Runner) setupEnvironEventHandler(ctx context.Context, plan *Plan, node 
 		env.LoadOutputVariables(node.inner.State.OutputVariables)
 	}
 
-	return execution.WithEnv(ctx, env)
+	return WithEnv(ctx, env)
 }
 
 func (r *Runner) execNode(ctx context.Context, node *Node) error {
@@ -1005,7 +1005,7 @@ func (r *Runner) shouldRepeatNode(ctx context.Context, node *Node, execErr error
 		return false
 	}
 
-	env := execution.GetEnv(ctx)
+	env := GetEnv(ctx)
 	shell := env.Shell(ctx)
 	switch rp.RepeatMode {
 	case core.RepeatModeWhile:
@@ -1013,9 +1013,9 @@ func (r *Runner) shouldRepeatNode(ctx context.Context, node *Node, execErr error
 		if rp.Condition != nil {
 			// Ensure node's own output variables are reloaded before evaluating the condition.
 			if node.inner.State.OutputVariables != nil {
-				env := execution.GetEnv(ctx)
+				env := GetEnv(ctx)
 				env.ForceLoadOutputVariables(node.inner.State.OutputVariables)
-				ctx = execution.WithEnv(ctx, env)
+				ctx = WithEnv(ctx, env)
 			}
 			err := EvalCondition(ctx, shell, rp.Condition)
 			return (err == nil) // Repeat if condition is met (no error)
@@ -1032,9 +1032,9 @@ func (r *Runner) shouldRepeatNode(ctx context.Context, node *Node, execErr error
 		if rp.Condition != nil {
 			// Ensure node's own output variables are reloaded before evaluating the condition.
 			if node.inner.State.OutputVariables != nil {
-				env := execution.GetEnv(ctx)
+				env := GetEnv(ctx)
 				env.ForceLoadOutputVariables(node.inner.State.OutputVariables)
-				ctx = execution.WithEnv(ctx, env)
+				ctx = WithEnv(ctx, env)
 			}
 			err := EvalCondition(ctx, shell, rp.Condition)
 			return (err != nil) // Repeat if condition is NOT met (error)
@@ -1088,8 +1088,8 @@ func calculateBackoffInterval(interval time.Duration, backoff float64, maxInterv
 	return interval
 }
 
-func NewEnv(ctx context.Context, step core.Step, plan *Plan) execution.Env {
-	env := execution.NewEnv(ctx, step)
+func NewPlanEnv(ctx context.Context, step core.Step, plan *Plan) Env {
+	env := NewEnvForStep(ctx, step)
 	for _, n := range plan.Nodes() {
 		if n.Step().ID != "" {
 			env.StepMap[n.Step().ID] = n.StepInfo()
